backend/requirements.txt
insanely-fast-whisper
flask
flask-cors
python-dotenv
